# artifacts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "共域 CoField API",
    "description": "以共同事件为核心的校园社交连接器。\n\n领域词汇见 CONTEXT.md；用户可见文案不使用这些词汇，映射见 docs/07。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/me/seeds": {
      "get": {
        "tags": [
          "seeds"
        ],
        "summary": "My Seeds",
        "description": "我的信箱。已经找到同行者的种子不再占地方。",
        "operationId": "my_seeds_api_me_seeds_get",
        "parameters": [
          {
            "name": "x-campus-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Campus-Id"
            }
          },
          {
            "name": "x-principal-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Principal-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SeedOut"
                  },
                  "title": "Response My Seeds Api Me Seeds Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/seeds/{intent_id}:respond": {
      "post": {
        "tags": [
          "seeds"
        ],
        "summary": "Respond",
        "description": "愿意参与，或者这次不感兴趣。\n\n**愿意不等于加入**：还要发起人挑中才成局。这一句必须在界面上说清楚，\n否则用户会以为自己已经答应了，然后在没被选中时觉得被放了鸽子。",
        "operationId": "respond_api_seeds__intent_id__respond_post",
        "parameters": [
          {
            "name": "intent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Intent Id"
            }
          },
          {
            "name": "x-campus-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Campus-Id"
            }
          },
          {
            "name": "x-principal-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Principal-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RespondRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeedOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/intents/{intent_id}/candidates": {
      "get": {
        "tags": [
          "seeds"
        ],
        "summary": "Candidates",
        "description": "谁说了愿意。发起人在这一屏挑人。",
        "operationId": "candidates_api_intents__intent_id__candidates_get",
        "parameters": [
          {
            "name": "intent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Intent Id"
            }
          },
          {
            "name": "x-campus-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Campus-Id"
            }
          },
          {
            "name": "x-principal-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Principal-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidatesOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/intents/{intent_id}/candidates/{who}:choose": {
      "post": {
        "tags": [
          "seeds"
        ],
        "summary": "Choose",
        "description": "挑一个人。收满就成局。\n\n**AI 不替发起人做这一下。** 它排序、它解释，但按下去的必须是人。",
        "operationId": "choose_api_intents__intent_id__candidates__who__choose_post",
        "parameters": [
          {
            "name": "intent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Intent Id"
            }
          },
          {
            "name": "who",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Who"
            }
          },
          {
            "name": "x-campus-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Campus-Id"
            }
          },
          {
            "name": "x-principal-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Principal-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidatesOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CandidateOut": {
        "properties": {
          "principal_id": {
            "type": "string",
            "format": "uuid",
            "title": "Principal Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "why": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Why"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "chosen": {
            "type": "boolean",
            "title": "Chosen",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "principal_id",
          "display_name",
          "why"
        ],
        "title": "CandidateOut",
        "description": "一个说了愿意的人。"
      },
      "CandidatesOut": {
        "properties": {
          "still_need": {
            "type": "integer",
            "title": "Still Need"
          },
          "chosen": {
            "items": {
              "$ref": "#/components/schemas/CandidateOut"
            },
            "type": "array",
            "title": "Chosen"
          },
          "willing": {
            "items": {
              "$ref": "#/components/schemas/CandidateOut"
            },
            "type": "array",
            "title": "Willing"
          },
          "still_thinking": {
            "type": "integer",
            "title": "Still Thinking"
          },
          "formed_event_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Formed Event Id"
          },
          "space_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Space Id"
          }
        },
        "type": "object",
        "required": [
          "still_need",
          "chosen",
          "willing",
          "still_thinking"
        ],
        "title": "CandidatesOut",
        "description": "发起人那一屏。\n\n**不给分数。** AI 排序并解释，但最后那一下必须是人点的。"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RespondRequest": {
        "properties": {
          "willing": {
            "type": "boolean",
            "title": "Willing"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "remind_me": {
            "type": "boolean",
            "title": "Remind Me",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "willing"
        ],
        "title": "RespondRequest"
      },
...
      "SeedOut": {
        "properties": {
          "intent_id": {
            "type": "string",
            "format": "uuid",
            "title": "Intent Id"
          },
          "goal": {
            "type": "string",
            "title": "Goal"
          },
          "said": {
            "type": "string",
            "title": "Said"
          },
          "needs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Needs"
          },
          "offers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Offers"
          },
          "when": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "When"
          },
          "where": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Where"
          },
          "team_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Min"
          },
          "team_max": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Max"
          },
          "from_name": {
            "type": "string",
            "title": "From Name"
          },
          "why": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Why"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "my_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "My Note"
          },
          "space_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Space Id"
          }
        },
        "type": "object",
        "required": [
          "intent_id",
          "goal",
          "said",
          "needs",
          "offers",
          "from_name",
          "why",
          "state"
        ],
        "title": "SeedOut",
        "description": "信箱里的一颗种子。\n\n**完整展示行动信息，有限展示发起人信息**（PRD 阶段三）：他要判断的是\n这件事值不值得参与，不是这个人够不够格。"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}